HW7_Kubeflow_AirFlow_pipelines/requirements.txt
# General dependencies
numpy>=1.26.0
pandas>=2.0.3
matplotlib>=3.8.0
scikit-learn>=1.4.0
Pillow>=10.0.0

# Нужно установить предварительно для решения проблем с компиляцией
PyYAML>=6.0.1

# Kubeflow dependencies
kfp==2.13.0  # Совместимая версия с кодом inference_pipeline.py
kubernetes>=28.1.0



# Airflow dependencies
apache-airflow>=2.8.1
apache-airflow-providers-docker>=3.8.0
apache-airflow-providers-cncf-kubernetes>=8.0.0

# ML dependencies
torch>=2.2.0
torchvision>=0.17.0

# Storage and data handling
google-cloud-storage>=2.13.0
boto3>=1.34.0
